+++ tests/Common/IntegrationTestCase.php
@@ -4,7 +4,6 @@ namespace DDTrace\Tests\Common;
 
 use PHPUnit\Framework\TestCase;
 
-
 /**
  * A basic class to be extended when testing integrations.
  */